Added Sequelize.
Initiated auto sync on startup.
Finished models setup.
Removed ES6.
