+++ config/sync.js
@@ -1,0 +1,8 @@
+const db = require('./connection')
+
+db.sync()
+  .then(() => {
+    console.log('Databases and Tables Created')
+  })
+
+
