--- src/app.js
@@ -1,50 +1,0 @@
-import createError from 'http-errors';
-import express from 'express';
-import path from 'path';
-import cookieParser from 'cookie-parser';
-import logger from 'morgan';
-import sqlite3 from 'sqlite3';
-import bodyParser from 'body-parser';
-
-import indexRouter from './routes/index';
-import loginRouter from './routes/login';
-import studentsRouter from './routes/students';
-import settingRouter from './routes/settings';
-import facultyRouter from './routes/faculty';
-
-const app = express();
-
-// view engine setup
-//app.set('views', path.join(__dirname, 'views'));
-//app.set('view engine', 'jade');
-
-app.use(logger('dev'));
-app.use(bodyParser.urlencoded({ extended: true }));
-app.use(bodyParser.json());
-app.use(cookieParser());
-app.use(express.static(path.join(__dirname, '../public')));
-
-app.use('/', indexRouter);
-app.use('/api/students', studentsRouter);
-app.use('/api/faculty', facultyRouter);
-app.use('/api/login', loginRouter);
-app.use('/api/settings', settingRouter);
-
-// catch 404 and forward to error handler
-app.use(function (req, res, next) {
-  next(createError(404));
-});
-
-// error handler
-app.use(function (err, req, res, next) {
-  // set locals, only providing error in development
-  res.locals.message = err.message;
-  res.locals.error = req.app.get('env') === 'development' ? err : {};
-
-  // render the error page
-  res.status(err.status || 500);
-  res.render('error');
-});
-
-export default app;
-
