--- src/models/examiner.js
@@ -1,6 +1,0 @@
-import Sequelize, { Model } from Sequelize;
-export default class Examiner extends Model { }
-
-Examiner.init({
-
-});
